new architecture for the other features
this branch is the new architecture i want to implement for preparing the next features.
in fact, at the begining i dosn't aspect to go so far on the project. So, the code was not really opti for other features.
by the features.py file, i range in classe the features and they are called in the main.py after the selection of wath the uesr want to.
the next features i whent to add is the exectution of program by only tutching your keyboard and whithout the help of a shell.
the shortcut and the .exe file will be editable by the user.
